# docs/swagger.yaml
swagger: "2.0"
info:
  description: "Space Time Machine is used to travel through the space and time."
  version: "1.0.0"
  title: "Space Time Machine"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "#"
basePath: "/v2"
schemes:
- "http"
paths:
  /traveller:
    post:
      tags:
      - "Traveller"
      summary: "Register as space traveller"
      description: ""
      operationId: "addTraveller"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Traveller object which has Name and email id"
        required: true
        schema:
          $ref: "#/definitions/TravellerInputForm"
      responses:
        "201":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Traveller"
        "409":
          description: "User already registered"
          
  /traveller/{pgi}:
    get:
      tags:
      - "Traveller"
      summary: "Find traveller by pgi"
      description: "Returns a single traveller"
      operationId: "getTravellerByPGI"
      produces:
      - "application/json"
      parameters:
      - name: "pgi"
        in: "path"
        description: "pgi of traveller to return"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Traveller"
        "404":
          description: "Traveller not found"
  /traveller/travel:
    post:
      tags:
      - "Space Travel"
      summary: "Perform space time travel"
      description: "Returns a single traveller Log"
      operationId: "doTravel"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Traveller object which has pgi, place and date"
        required: true
        schema:
          $ref: "#/definitions/TravelLog"
      responses:
        "201":
          description: "successful operation"
          schema:
            $ref: "#/definitions/TravelLog"
        "409":
          description: "Other version of yourself is already there. Dont want to create Paradox"
          
  /traveller/travelHistory/{pgi}:
    get:
      tags:
      - "Space Travel"
      summary: "Find travel logs by pgi"
      description: "Returns a list of travelLog"
      operationId: "getTravelLogByPGI"
      produces:
      - "application/json"
      parameters:
      - name: "pgi"
        in: "path"
        description: "pgi of traveller"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/TravelLogs"
        "404":
          description: "TravelLog not found"
definitions:
  Traveller:
    type: "object"
    properties:
      pgi:
        type: "string"
      name:
        type: "string"
      email:
        type: "string"
  TravelLog:
    type: "object"
    properties:
      pgi:
        type: "string"
      place:
        type: "string"
      date:
        type: "string"
        pattern: "yyyy-MM-dd"
  TravellerInputForm:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
  TravelLogs:
    type: "array"
    items: 
      $ref: "#/definitions/TravelLog"

externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"